+++ deploy/Invoke-SmokeTests.ps1
@@ -1,9 +1,15 @@
 # These tests are runs as part of the deployment process to ensure the newly published module is working as expected.
 # These tests run against the installed module, not the source code, so they do not use mocks and are more of a real-world test.
 # Since mocks are not used, we must be careful to not rely on state stored on the machine, such as the module Configuration file.
+# To run these tests on your local machine, see the comments in the BeforeAll block.
 
 BeforeAll {
 	Import-Module -Name tiPS -Force
+
+	# To run these tests on your local machine, comment out the Import-Module command above and uncomment the one below.
+	# 	Do this to use the module version from source code, not the installed version.
+	# 	This is necessary to test functionality that you've added to the module, but have not yet published and installed.
+	# Import-Module "$PSScriptRoot\..\src\tiPS" -Force
 }
 
 Describe 'Get-PowerShellTip' {
@@ -33,6 +39,12 @@ Describe 'Get-PowerShellTip' {
 			$allTips = Get-PowerShellTip -All
 			$allTips | Should -Not -BeNullOrEmpty
 			$allTips.Count | Should -BeGreaterThan 2
+		}
+
+		It 'Should not return any expired tips, as they should have been removed during initialization' {
+			$allTips = Get-PowerShellTip -All
+			$expiredTips = $allTips.Values | Where-Object { $_.ExpiryDate -lt [DateTime]::Today }
+			$expiredTips | Should -BeNullOrEmpty
 		}
 	}
 }